+++ modules/html-storybook/src/Multiselect.stories.tsx
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import { Multiselect } from '@reykjavik/hanna-react/Multiselect';
+import { boolean } from '@storybook/addon-knobs';
 
+import { HiddenTiger } from './utils/HiddenTiger.js';
+import { getFormFieldKnobs } from './utils/knobs.js';
 import { StoryComponent, StoryParameters } from './storytypes.js';
 
 // ===========================================================================
@@ -14,41 +17,61 @@ export default {
 };
 
 const options = [
-  { value: 'chocolate', label: 'Chocolate' },
-  { value: 'strawberry', label: 'Strawberry' },
-  { value: 'vanilla', label: 'Vanilla' },
-  { value: 'coffee', label: 'Coffee' },
-  { value: 'mint-chocolate-chip', label: 'Mint Chocolate Chip' },
-  { value: 'rocky-road', label: 'Rocky Road' },
-  { value: 'cookies-and-cream', label: 'Cookies and Cream' },
-  { value: 'butter-pecan', label: 'Butter Pecan' },
-  { value: 'pistachio', label: 'Pistachio' },
-  { value: 'maple-walnut', label: 'Maple Walnut' },
-  { value: 'caramel', label: 'Caramel' },
-  { value: 'fudge', label: 'Fudge' },
+  'Chocolate',
+  'Strawberry',
+  'Vanilla',
+  'Coffee',
+  'Mint Chocolate Chip',
+  'Rocky Road',
+  'Cookies and Cream',
+  'Butter Pecan',
+  'Pistachio',
+  'Maple Walnut',
+  'Caramel',
+  'Fudge',
 ];
 
 export const _Multiselect: StoryComponent = () => {
+  const nowrap = boolean('No-wrap', false);
+  const ffProps = getFormFieldKnobs({ hideLabel: false });
+
+  const empty = boolean('Empty', true);
+  const forceSearchable = boolean('Searchable options', false);
+
+  const values = empty ? [] : nowrap ? options : options.slice(2, 4);
+
+  const key = [nowrap, empty, forceSearchable].join('-') + JSON.stringify(ffProps);
+
   return (
-    <>
-      <Multiselect items={options} label="Veldu tegund" />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-      <br />
-    </>
+    <Fragment key={key}>
+      <Multiselect
+        {...ffProps}
+        label="Veldu tegund"
+        name="flavours"
+        options={options}
+        defaultValue={values}
+        forceSearchable={forceSearchable}
+      />
+
+      <HiddenTiger>
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+        <br />
+      </HiddenTiger>
+    </Fragment>
   );
 };